requirements.txt
pystray~=0.19.5
2fas
pyperclip~=1.9.0
cairosvg~=2.7.1
lib2fas~=0.1.7
pyjson5~=1.6.6
pillow~=10.4.0
tray_manager
# Notificaciones nativas para macOS
pync; sys_platform == 'darwin'
# PyObjC para integración con macOS
pyobjc-core; sys_platform == 'darwin'
pyobjc-framework-Cocoa; sys_platform == 'darwin'
